% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.cal.conv.enviro.multi.out.R
\name{def.cal.conv.enviro.multi.out}
\alias{def.cal.conv.enviro.multi.out}
\title{Test function for producing multiple outputs, including uncertainty info and cal flags, for the EnviroSCAN sensor}
\usage{
def.cal.conv.enviro.multi.out(
  data = data.frame(data = base::numeric(0)),
  varConv = setdiff(base::names(data), c("source_id", "site_id", "readout_time")),
  calSlct = NULL,
  Meta = list(),
  log = NULL
)
}
\arguments{
\item{data}{Data frame of raw, uncalibrated measurements. This data frame must have a column
called "readout_time"}

\item{varConv}{A character array of the target variables (columns) in the data frame \code{data} for 
which calibrated output will be computed (all other columns will be ignored). Defaults to 
all column names in \code{data} except 'source_id','site_id','readout_time'.}

\item{calSlct}{A named list of data frames, each list element corresponding to a 
variable (column) to calibrate. The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, 
as returned from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{Meta}{(Optional) List object containing additional metadata for use in 
this function as needed. Defaults to an empty list, but this example requires that the list
item Meta$Locations is input to work properly.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A named list of outputs: \cr
data: The input data frame 'data' transformed by this function. \cr
ucrtData: A named list of data frames providing uncertainty data for each output variable in data \cr
ucrtCoef: A named list of data frames providing uncertainty coefficents for each output variable in data
}
\description{
Definition function. Template for custom calibration function that produces the following:\cr
- calibrated data with multiple output variables per input variable
- uncertainty estimates for all calibrated outputs
- uncertainty coefficients for all calibrated outputs
- calibration quality flags for all calibrated outputs (not implemented yet, but possible)
- sensor depths
}
\examples{
None currently
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Calibrated Measurements and Level 1 Data Products Uncertainty Budget Plan
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.conv.poly.b}

\link[NEONprocIS.cal]{def.cal.conv.poly.m}

\link[NEONprocIS.cal]{wrap.cal.conv.dp0p}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
Edward Ayres \email{eayres@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
