% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.cal.conv.poly.h.R
\name{def.cal.conv.poly.h}
\alias{def.cal.conv.poly.h}
\title{Convert raw to calibrated data using NEON CVALH polynomial calibration coefficients}
\usage{
def.cal.conv.poly.h(
  data = data.frame(data = base::numeric(0)),
  varConv = base::names(data)[1],
  calSlct = NULL,
  Meta = list(),
  log = NULL
)
}
\arguments{
\item{data}{Data frame of raw, uncalibrated measurements. This data frame must have a column
called "readout_time" with POSIXct timestamps}

\item{varConv}{A character string of the target variables (columns) in the data frame \code{data} for 
which calibrated output will be computed (all other columns will be ignored). Defaults to the first
column in \code{data}.}

\item{calSlct}{A named list of data frames, each list element corresponding to a 
variable (column) to calibrate. The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, 
as returned from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{Meta}{Unused in this function. Defaults to an empty list. See the inputs to 
NEONprocIS.cal::wrap.cal.conv.dp0p for what this input is.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
The input data frame, with the columns specified in input \code{varConv} updated with 
calibrations applied.
}
\description{
Definition function. Apply NEON calibration polynomial function contained in coefficients 
CVALH0, CVALH1, CVALH2, etc. to convert raw data to calibrated data.
}
\examples{
Not Run
data=data.frame(readout_time=as.POSIXct('2025-01-01','2025-01-02','2025-01-03'),var1=c(1,2,3),var2=c(4,5,6))
calSlct <- NEONprocIS.cal::wrap.cal.slct(
               DirCal = '/path/to/calibration/files',
               NameVarExpc = c('var1','var2'),
               TimeBgn = as.POSIXct('2025-01-01'),
               TimeEnd = as.POSIXct('2025-01-04'),
               )
dataCal <- def.cal.conv.poly(data=data,varConv=c('var1','var2'),calSlct=calSlct)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Calibrated Measurements and Level 1 Data Products Uncertainty Budget Plan
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.conv.poly}

\link[NEONprocIS.cal]{def.cal.conv.poly.m}

\link[NEONprocIS.cal]{wrap.cal.conv.dp0p}
}
\author{
Nora Catolico \email{ncatolico@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
