% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.cal.conv.swc.test.R
\name{def.cal.conv.swc.test}
\alias{def.cal.conv.swc.test}
\title{Test function for producing multiple calibrated outputs for each calibrated SWC stream}
\usage{
def.cal.conv.swc.test(
  data = data.frame(data = base::numeric(0)),
  varConv = base::names(data)[1],
  calSlct = NULL,
  Meta = list(),
  log = NULL
)
}
\arguments{
\item{data}{Data frame of raw, uncalibrated measurements. This data frame must have a column
called "readout_time"}

\item{varConv}{A character array of the target variables (columns) in the data frame \code{data} for 
which calibrated output will be computed (all other columns will be ignored). Defaults to the first
column in \code{data}.}

\item{calSlct}{A named list of data frames, each list element corresponding to a 
variable (column) to calibrate. The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, 
as returned from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{Meta}{(Optional) List object containing additional metadata for use in 
this function as needed. Defaults to an empty list.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
The input data frame 'data' transformed by this function.
}
\description{
Definition function. Apply NEON calibration polynomial function contained in coefficients 
CVALA0, CVALA1, CVALA2, etc. to convert raw data to calibrated data.
}
\examples{
data=data.frame(data=c(1,2,3))
infoCal <- data.frame(Name=c('CVALA1','CVALA0'),Value=c(10,1),stringsAsFactors=FALSE)
def.cal.conv.poly(data=data,infoCal=infoCal)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Calibrated Measurements and Level 1 Data Products Uncertainty Budget Plan
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.conv.poly.b}

\link[NEONprocIS.cal]{def.cal.conv.poly.m}

\link[NEONprocIS.cal]{wrap.cal.conv.dp0p}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
