% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.cal.func.poly.R
\name{def.cal.func.poly}
\alias{def.cal.func.poly}
\title{Create calibration polynomial function from NEON CVAL coefficients}
\usage{
def.cal.func.poly(infoCal, Prfx = "CVALA", log = NULL)
}
\arguments{
\item{infoCal}{List of calibration and uncertainty information read from a NEON calibration file
(as from NEONprocIS.cal::def.read.cal.xml). Included in this list must be infoCal$cal, which is
a data frame of uncertainty coefficents. Columns of this data frame are:\cr
\code{Name} String. The name of the coefficient. \cr
\code{Value} String or numeric. Coefficient value. Will be converted to numeric. \cr}

\item{Prfx}{String. The prefix for the polynomial coefficients. Default is 'CVALA', indicating 
that the polynomial coefficients will follow the convention CVALA0, CVALA1, CVALA2, etc., where
the number indicates the polynomial power CVALA0 + CVALA1x + CVALA2x^2...}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A polynomial (model) object of the polynomial calibration function
}
\description{
Definition function. Creates a polynomical object of the calibration function from NEON 
calibration coefficients (from e.g. NEONprocIS.cal::def.read.cal.xml).
}
\examples{
infoCal <- list(cal=data.frame(Name=c('CVALA1','CVALA0'),Value=c(10,1),stringsAsFactors=FALSE))
def.cal.func.poly(infoCal=infoCal)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.conv.poly}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
