% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.cal.meta.R
\name{def.cal.meta}
\alias{def.cal.meta}
\title{Compile metadata for calibrations}
\usage{
def.cal.meta(fileCal, log = NULL)
}
\arguments{
\item{fileCal}{Character vector of the full or relative paths to each calibration file}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of calibration metadata
}
\description{
Definition function. Read in any number of calibration files and compile their metadata, including
file name, certificate number, and valid date range valid date range.
}
\examples{
# Not run
# fileCal <- c('/path/to/file1.xml','/path/to/file2.xml')
# metaCal <- NEONprocIS.cal::def.cal.meta(fileCal=fileCal)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Calibrated Measurements and Level 1 Data Products Uncertainty Budget Plan
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
