% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.cal.slct.R
\name{def.cal.slct}
\alias{def.cal.slct}
\title{Determine applicable date ranges for calibration}
\usage{
def.cal.slct(metaCal = NULL, TimeBgn, TimeEnd, TimeExpiMax = NULL,
  log = NULL)
}
\arguments{
\item{metaCal}{A data frame of calibration metadata as returned from NEONprocIS.cal::def.cal.meta}

\item{TimeBgn}{A POSIXct timestamp of the start date of interest (inclusive)}

\item{TimeEnd}{A POSIXct timestamp of the end date of interest (exclusive)}

\item{TimeExpiMax}{A difftime object of the maxumum time since expiration for which an expired
calibration may be used (e.g. TimeExpiMax <- as.difftime(10,units='days'). Defaults to NULL,
which allows an expired calibration to always be used.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of \cr
\code{timeBgn} POSIXct. The start date-time (inclusive) for which the calibration applies \cr
\code{timeEnd} POSIXct. The end date-time (exclusive) for which the calibration applies \cr
\code{file} Character. The name of the calibration file referring to metaCal$file that applies for the date range in columns timeBgn and timeEnd \cr
\code{id} Numeric. The calibration id referring to metaCal$id that applies for the date range in columns timeBgn and timeEnd \cr
\code{expi} Boolean (TRUE/FALSE). TRUE if the best available calibration is expired for the time period
}
\description{
Definition function. Given calibration information from several calibration files (as returned
from NEONprocIS.cal::def.cal.met), and a maximum number of days allowed to use expired
calibrations after their expiration date, determine the calibrations that apply (and their
application time ranges) for a date range of interest. Calibration valid date ranges and
calibration certificate numbers (ID) determine the most relevant calibration to apply, following
this choice order (1 chosen first):
   1. higher ID & date of interest within valid date range
   2. lower ID & date of interest within valid date range
   3. expired cal with nearest valid end date to beginning date of interest
   4. lower ID if expiration dates equal for multipe cals in #3
Note that calibrations with a valid date range beginning after the date range of interest and
calibrations that are expired more than their max allowable days since expiration are treated
as if they don't exist. These time periods in the output will be filled with NA values for the 
calibration metadata with the expired flag = TRUE.
}
\examples{
# Not run
# fileCal <- c('/path/to/file1.xml','/path/to/file2.xml')
# metaCal <- NEONprocIS.cal::def.cal.meta(fileCal=fileCal)
# TimeBgn <- base::as.POSIXct('2019-01-01',tz='GMT')
# TimeEnd <- base::as.POSIXct('2019-01-02',tz='GMT')
# TimeExpiMax <- base::as.difftime(30,units='days') # allow cals to be used up to 30 days after expiration
# NEONprocIS.cal::def.cal.slct(metaCal=metaCal,TimeBgn=timeBgn,TimeEnd=timeEnd)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.cal]{def.cal.meta}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{currently}
\keyword{none}
