% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.cal.susp.R
\name{def.qf.cal.susp}
\alias{def.qf.cal.susp}
\title{Quality flag for suspect calibration}
\usage{
def.qf.cal.susp(data = base::numeric(0), infoCal = NULL, log = NULL)
}
\arguments{
\item{data}{Numeric vector of raw measurements}

\item{infoCal}{A list of calibration information as returned from NEONprocIS.cal::def.read.cal.xml. 
One list element must be \code{cal}, which is a data frame of polynomial calibration coefficients. 
This data frame must include columns:\cr
\code{Name} String. The name of the coefficient.\cr
\code{Value} String or numeric. Coefficient value. Will be converted to numeric. \cr
Defaults to NULL, in which case flag value will be -1.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A numeric vector the same length as input argument "data" of suspect calibration flag 
values (1=bad,0=good,-1=could not evaluate).
}
\description{
Definition function. Check the CVALR1 coefficient in the calibration file. If it is present,
set the suspect calibration flag for all data values to the value of the coefficient 
(1=bad,0=good). If the CVALR1 coefficient is not present, set the suspect calibration flag 
to 0. If no calibration information is available, set the flag to -1.
}
\examples{
data <- c(1,2,3,4,5)
infoCal <- list(cal=data.frame(Name=c('CVALR1','CVALA0'),Value=c(1,0),stringsAsFactors=FALSE))
qfSusp <- NEONprocIS.cal::def.qf.cal.susp(data,infoCal)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
