% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.cal.xml.R
\name{def.read.cal.xml}
\alias{def.read.cal.xml}
\title{Read NEON calibration XML file}
\usage{
def.read.cal.xml(NameFile, Vrbs = TRUE)
}
\arguments{
\item{NameFile}{String. Name (including relative or absolute path) of calibration file.}

\item{Vrbs}{(Optional) Logical. If TRUE, returns the full contents of the calibration file as an additional output.}
}
\value{
A named list:\cr
timeVali = a list of POSIXct valid start & end date-times for the calibration 
cal = a data frame of calibration coefficients
ucrt = a data frame of uncertainty coefficients
file = a list of the contents of the full calibration file. Only returned if Vrbs = TRUE.
}
\description{
Definition function. Read in a NEON calibration XML file.
}
\examples{
Currently none
}
\references{
Currently none
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
