% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.ucrt.coef.fdas.R
\name{def.read.ucrt.coef.fdas}
\alias{def.read.ucrt.coef.fdas}
\title{Read NEON FDAS uncertainty coefficients}
\usage{
def.read.ucrt.coef.fdas(NameFile, log = NULL)
}
\arguments{
\item{NameFile}{String. Name (including relative or absolute path) of json file.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of FDAS uncertainty coefficients:\cr
\code{Name} Character. Name of the coefficient.\cr
\code{Value} Character. Value of the coefficient.\cr
\code{.attrs} Character. Relevant attribute (i.e. units)\cr
}
\description{
Definition function. Read in NEON FDAS uncertainty coefficients
}
\examples{
Currently none
}
\references{
Currently none
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
