% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.comb.R
\name{def.ucrt.comb}
\alias{def.ucrt.comb}
\title{Combine measurement uncertainties by adding them in quadrature.}
\usage{
def.ucrt.comb(ucrt, log = NULL)
}
\arguments{
\item{ucrt}{Numeric data frame of uncertainties to be combined.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A data frame with a single numeric column:\cr
\code{ucrtComb} - combined measurement uncertainty resulting by adding in quadrature all the
uncertainties provided in the input data frame \code{ucrt}.
}
\description{
Definition function. Reads in a data frame of uncertainties and adds each row in quadrature.
NA values are NOT ignored.
}
\examples{
ucrt <- data.frame(ucrtA=c(1,2,1,1,2),ucrtB=c(5,6,7,8,9),stringsAsFactors=FALSE)
NEONprocIS.cal::def.ucrt.comb(data=data,ucrtCoef=ucrtCoef)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.ucrt.fdas.rstc}

\link[NEONprocIS.cal]{def.ucrt.fdas.volt}

\link[NEONprocIS.cal]{def.ucrt.meas}

\link[NEONprocIS.base]{def.log.init}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{combined}
\keyword{uncertainty}
