% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.expn.R
\name{def.ucrt.expn}
\alias{def.ucrt.expn}
\title{Compute expanded measurement uncertainty (95 percent confidence)}
\usage{
def.ucrt.expn(ucrtComb, log = NULL)
}
\arguments{
\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}

\item{ucrt}{Numeric data vector of combined measurement uncertainty (1 sigma)}
}
\value{
A data frame with a single numeric column: \cr
\code{ucrtExpn} - expanded measurement uncertainty
}
\description{
Definition function. Compute expanded measurement uncertainty at 95 percent confidence from combined
measurement uncertainty
}
\examples{
ucrtComb <- c(1,2,1,1,2)
NEONprocIS.cal::def.ucrt.expn(ucrtComb=ucrtComb)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.ucrt.fdas.rstc.poly}

\link[NEONprocIS.cal]{def.ucrt.fdas.volt.poly}

\link[NEONprocIS.cal]{def.ucrt.meas.cnst}

\link[NEONprocIS.cal]{def.ucrt.meas.mult}

\link[NEONprocIS.cal]{def.ucrt.comb}

\link[NEONprocIS.base]{def.log.init}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{combined}
\keyword{uncertainty}
