% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.fdas.rstc.poly.R
\name{def.ucrt.fdas.rstc.poly}
\alias{def.ucrt.fdas.rstc.poly}
\title{Compute uncertainty attributed to NEON FDAS resistance measurements (and CVALA polynomial calibration conversion)}
\usage{
def.ucrt.fdas.rstc.poly(
  data = data.frame(data = base::numeric(0)),
  varUcrt = base::names(data)[1],
  calSlct = NULL,
  Meta = list(),
  log = NULL
)
}
\arguments{
\item{data}{Data frame of raw measurements. Note that the column/variable indicated in varUcrt must be numeric.}

\item{varUcrt}{A character string of the target variable (column) in the data frame \code{data} for 
which FDAS uncertainty data will be computed (all other columns will be ignored). Note that for other
uncertainty functions this variable may not need to be in the input data frame, so long as the function
knows that. Defaults to the first column in \code{data}.}

\item{calSlct}{A named list of data frames, list element corresponding to the variables in
varUcrt. The data frame in each list element holds information about the calibration files and 
time periods that apply to the variable, as returned from NEONprocIS.cal::def.cal.slct. 
See documentation for that function. Assign NULL to list elements (variables) for which calibration
information is not applicable (i.e. a function other than def.ucrt.meas.cnst is used to compute its
uncertainty).}

\item{Meta}{Named list of metadata for use in this function. Meta is required to contain
list element ucrtCoefFdas, which is a data frame of FDAS uncertainty coefficients, as read by 
NEONprocIS.cal::def.read.ucrt.coef.fdas. Columns include:\cr
\code{Name} Character. Name of the coefficient.\cr
\code{Value} Character. Value of the coefficient.\cr
\code{.attrs} Character. Relevant attribute (i.e. units)\cr}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A named list of data frames, each list element corresponding to a variable specified in 
\code{varUcrt}. Each data frame contains the following data columns:\cr
\code{ucrt$raw} - raw reading value (i.e. same as input data)\cr
\code{ucrt$dervCal} - 1st derivative of calibration function evaluated at raw reading value (e.g.
partial derivative of a temperature measurement with respect to the resistance reading)\cr
\code{ucrt$ucrtFdas} - standard uncertainty of individual measurement introduced by the Field DAS \cr
}
\description{
Definition function. Computes L0' uncertainty imposed by a NEON field data acquisition system (FDAS)
when making resistance measurements and using a polynomial calibration conversion function with CVALA 
coefficients.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Calibrated Measurements and Level 1 Data Products Uncertainty Budget Plan
}
\seealso{
\link[NEONprocIS.cal]{def.cal.func.poly}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{wrap.ucrt.dp0p}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
