% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.fdas.volt.poly.R
\name{def.ucrt.fdas.volt.poly}
\alias{def.ucrt.fdas.volt.poly}
\title{Compute uncertainty attributed to NEON FDAS voltage measurements (and CVALA polynomial calibration conversion)}
\usage{
def.ucrt.fdas.volt.poly(
  data = data.frame(data = base::numeric(0)),
  infoCal = NULL,
  varUcrt = base::names(data)[1],
  calSlct = NULL,
  log = NULL
)
}
\arguments{
\item{data}{Numeric data frame of raw measurements.}

\item{infoCal}{List of calibration and uncertainty information read from a NEON calibration file
(as from NEONprocIS.cal::def.read.cal.xml). Included in this list must be infoCal$cal and info$ucrt,
which are data frames of calibration coefficients and uncertainty coeffcients, respectively.
Columns of these data frames are:\cr
\code{Name} String. The name of the coefficient. \cr
\code{Value} String or numeric. Coefficient value. Will be converted to numeric. \cr}

\item{varUcrt}{A character string of the target variable (column) in the data frame \code{data} for 
which FDAS uncertainty data will be computed (all other columns will be ignored). Note that for other
uncertainty functions this variable may not need to be in the input data frame, so long as the function
knows that. Defaults to the first column in \code{data}.}

\item{calSlct}{Unused in this function. Defaults to NULL. See the inputs to 
NEONprocIS.cal::wrap.ucrt.dp0p for what this input is.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A data frame with the following variables:\cr
\code{ucrt$raw} - raw reading value (i.e. same as input data)\cr
\code{ucrt$dervCal} - 1st derivative of calibration function evaluated at raw reading value (e.g. partial derivative
of a radiation measurement with respect to the voltage reading)\cr
\code{ucrt$ucrtFdas} - standard uncertainty of individual measurement introduced by the Field DAS \cr
}
\description{
Definition function. Computes L0' uncertainty imposed by a NEON field data acquisition system (FDAS)
when making voltage measurements and using a polynomial calibration conversion function with CVALA 
coefficients.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Calibrated Measurements and Level 1 Data Products Uncertainty Budget Plan
}
\seealso{
\link[NEONprocIS.cal]{def.cal.func.poly}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{wrap.ucrt.dp0p}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
