% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.fdas.volt.R
\name{def.ucrt.fdas.volt}
\alias{def.ucrt.fdas.volt}
\title{Compute uncertainty attributed to NEON FDAS voltage measurements}
\usage{
def.ucrt.fdas.volt(data = base::numeric(0), infoCal = NULL, log = NULL)
}
\arguments{
\item{data}{Numeric vector of raw voltage measurements}

\item{infoCal}{List of calibration and uncertainty information read from a NEON calibration file
(as from NEONprocIS.cal::def.read.cal.xml). Included in this list must be infoCal$cal and info$ucrt,
which are data frames of calibration coefficients and uncertainty coeffcients, respectively.
Columns of these data frames are:\cr
\code{Name} String. The name of the coefficient. \cr
\code{Value} String or numeric. Coefficient value. Will be converted to numeric. \cr}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A data frame with the following variables:\cr
\code{ucrt$raw} - raw reading value (i.e. same as input data)\cr
\code{ucrt$dervCal} - 1st derivative of calibration function evaluated at raw reading value (e.g. partial derivative
of a radiation measurement with respect to the voltage reading)\cr
\code{ucrt$ucrtFdas} - standard uncertainty of individual measurement introduced by the Field DAS \cr
}
\description{
Definition function. Computes uncertainty imposed by a NEON field data acquisition system (FDAS)
when making voltage measurements.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Calibrated Measurements and Level 1 Data Products Uncertainty Budget Plan
}
\seealso{
\link[NEONprocIS.cal]{def.cal.func.poly}

\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
