% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.meas.cnst.R
\name{def.ucrt.meas.cnst}
\alias{def.ucrt.meas.cnst}
\title{Compute individual measurement calibration uncertainty as a constant from NEON CVAL coefficient U_CVALA1}
\usage{
def.ucrt.meas.cnst(
  data = data.frame(data = base::numeric(0)),
  infoCal = NULL,
  varUcrt = base::names(data)[1],
  calSlct = NULL,
  log = NULL
)
}
\arguments{
\item{data}{Numeric data frame of raw measurements.}

\item{infoCal}{List of calibration and uncertainty information read from a NEON calibration file
(as from NEONprocIS.cal::def.read.cal.xml). Included in this list must be infoCal$ucrt, which is
a data frame of uncertainty coefficents. Columns of this data frame are:\cr
\code{Name} String. The name of the coefficient. \cr
\code{Value} String or numeric. Coefficient value. Will be converted to numeric. \cr}

\item{varUcrt}{A character string of the target variable (column) in the data frame \code{data} for 
which uncertainty data will be computed (all other columns will be ignored). Note that for other
uncertainty functions this variable may not need to be in the input data frame, so long as the function
knows that. Defaults to the first column in \code{data}.}

\item{calSlct}{Unused in this function. Defaults to NULL. See the inputs to 
NEONprocIS.cal::wrap.ucrt.dp0p for what this input is.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A data frame with the following variables:\cr
\code{ucrtMeas} - combined measurement uncertainty for an individual reading. Includes the
repeatability and reproducibility of the sensor and the lab DAS and uncertainty of the
calibration procedures and coefficients including uncertainty in the standard (truth).
}
\description{
Definition function. Accepts a L0 data vector and NEON uncertainty information as produced
by NEONprocIS.cal::def.read.cal.xml and returns a vector of individual measurement
uncertainties for each data value. The uncertainty computed is simply a constant value 
represented by NEON calibration coefficient U_CVALA1.
}
\examples{
data <- data.frame(data=c(1,6,7,0,10))
infoCal <- list(ucrt = data.frame(Name=c('U_CVALA1','U_CVALA3'),Value=c(0.1,5),stringsAsFactors=FALSE))
def.ucrt.meas.cnst(data=data,infoCal=infoCal)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.ucrt.fdas.rstc.poly}

\link[NEONprocIS.cal]{def.ucrt.fdas.volt.poly}

\link[NEONprocIS.base]{def.log.init}

\link[NEONprocIS.cal]{wrap.ucrt.dp0p}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{calibration,}
\keyword{uncertainty}
