% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.meas.rh.dew.frst.pt.R
\name{def.ucrt.meas.rh.dew.frst.pt}
\alias{def.ucrt.meas.rh.dew.frst.pt}
\title{Uncertainty for dew/frost point from the relative humidity sensor as part of the relative
humidity transition}
\usage{
def.ucrt.meas.rh.dew.frst.pt(
  data = data.frame(data = base::numeric(0)),
  varUcrt = "dew_point",
  calSlct = NULL,
  Meta = list(),
  log = NULL
)
}
\arguments{
\item{data}{Numeric data frame of raw measurements of Temperature, relative humidity, 
and dew/frost point from the relative humidity sensor. Must include POSIXt variable readout_time.}

\item{varUcrt}{A character string of the target variable (column) in the data frame \code{data} for 
which uncertainty output will be computed (all other columns will be ignored). Defaults to "dew_point"}

\item{calSlct}{A named list of data frames, list element corresponding to the variables in
needed in this function (temperature, relative_humidity). The data frame in each list 
element holds information about the calibration files and 
time periods that apply to the variable, as returned from NEONprocIS.cal::def.cal.slct. 
See documentation for that function. Assign NULL to list elements (variables) for which calibration
information is not applicable (i.e. a function other than def.ucrt.meas.cnst is used to compute its
uncertainty).}

\item{Meta}{Unused in this function. Defaults to an empty list. See the inputs to 
NEONprocIS.cal::wrap.ucrt.dp0p for what this input is.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A named list containing a data frame, the list element named for the variable for which uncertainty 
estimates are computed (matching varUcrt). The uncertainty data frame contains the following variables:\cr
\code{ucrtMeas} - combined measurement uncertainty for an individual dew/frost point reading. Includes the
repeatability and reproducibility of the sensor and the lab DAS and uncertainty of the
calibration procedures and coefficients including uncertainty in the standard (truth).
\code{ucrt_dfpt_t_L1} - partial uncertainty (degrees C) of individual dew/frost point
temperature measurements with respect to ambient temperature
\code{ucrt_dfpt_rh_L1} - partial uncertainty (degrees C) of individual dew/frost point temperature 
measurements with respect to ambient relative humidity for Level 1 Mean uncertainty calculations
}
\description{
Definition function. Alternative calibration uncertainty function. Compute  
measurement uncertainty based off of the L0 temperature, relative humidity, and dew/frost point data values from 
the relative humidity sensor according to NEON.DOC.000851 - NEON Algorithm Theoretical Basis   
Document (ATBD): Humidity and Temperature Sensor.
}
\examples{
data <- data.frame(dew_point=c(1,-1,5,4,4.5),temperature=c(2,3,6,8,5),relative_humidity=c(1,6,7,0,10),
readout_time=as.POSIXct(c('2019-01-01 02:00','2019-01-01 04:01','2019-01-01 06:02','2019-01-01 08:01','2019-01-01 10:02'),tz='GMT'))
calSlct=list("temperature"= data.frame(timeBgn=as.POSIXct("2019-01-01",tz="GMT"),
timeEnd=as.POSIXct("2019-01-02",tz="GMT"),file = "30000000000080_WO29705_157555.xml",id = 157555, expi= FALSE),
"relative_humidity"= data.frame(timeBgn=as.POSIXct("2019-01-01",tz="GMT"),
timeEnd=as.POSIXct("2019-01-02",tz="GMT"),file = "30000000000080_WO29705_157554.xml",id = 157554, expi= FALSE),
"dew_point"= data.frame(timeBgn=as.POSIXct("2019-01-01",tz="GMT"),timeEnd=as.POSIXct("2019-01-02",tz="GMT"),
file = "30000000000080_WO29705_157556.xml",id = 157556, expi= FALSE))
def.ucrt.meas.rh.dew.frst.pt(data=data,calSlct=calSlct)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
NEON.DOC.000851 NEON Algorithm Theoretical Basis Document (ATBD): Humidity and Temperature Sensor
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.ucrt.fdas.rstc.poly}

\link[NEONprocIS.cal]{def.ucrt.fdas.volt.poly}

\link[NEONprocIS.base]{def.log.init}

\link[NEONprocIS.cal]{wrap.ucrt.dp0p}
}
\author{
Edward Ayres \email{eayres@battelleecology.org}
}
\keyword{L0',}
\keyword{calibration,}
\keyword{dew}
\keyword{frost}
\keyword{hmp155,}
\keyword{point}
\keyword{point,}
\keyword{uncertainty,}
