% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.meas.swc.test.R
\name{def.ucrt.meas.swc.test}
\alias{def.ucrt.meas.swc.test}
\title{Compute standard and alternate measurement calibration uncertainties for SWC measurements}
\usage{
def.ucrt.meas.swc.test(
  data = data.frame(data = base::numeric(0)),
  varUcrt = base::names(data)[1],
  calSlct = NULL,
  Meta = list(),
  log = NULL
)
}
\arguments{
\item{data}{Numeric data frame of raw measurements.}

\item{varUcrt}{A character array of the target variables (columns) in the data frame \code{data} for 
which uncertainty output will be computed (all other columns will be ignored). Defaults to the first
column in \code{data}.}

\item{calSlct}{Unused in this function. Defaults to NULL. See the inputs to 
NEONprocIS.cal::wrap.ucrt.dp0p for what this input is.}

\item{Meta}{Unused in this function. Defaults to an empty list. See the inputs to 
NEONprocIS.cal::wrap.ucrt.dp0p for what this input is.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A named list of data frames, each list named for the variable for which uncertainty 
estimates are contained within. The standard computations are provided in list elements named for
each variable provided in \code{varUcrt}, and the alternate computations provided in list elements 
named for each variable provided in \code{varUcrt} with "Alt" appended to the name. 
Each uncertainty data frame contains the following variables:\cr
\code{ucrtMeas} - combined measurement uncertainty for an individual reading. Includes the
repeatability and reproducibility of the sensor and the lab DAS and uncertainty of the
calibration procedures and coefficients including uncertainty in the standard (truth).
}
\description{
Definition function. Accepts L0 data and NEON uncertainty information as produced
by NEONprocIS.cal::def.read.cal.xml and returns a list of data frames of individual measurement
uncertainties for each data value. Computes the standard CVAL uncertainty as the L0 value 
multipled by NEON calibration uncertainty coefficient U_CVALA1. Also computes an alternate 
uncertainty data frame computed is the L0 value multipled by NEON calibration uncertainty 
coefficient U_CVALD1 minus 10.
}
\examples{
# Not run
data <- data.frame(data=c(1,6,7,0,10))
calSlct <- NEONprocIS.cal::wrap.cal.slct(
  DirCal = "/path/to/datum/calibration/folder",
  NameVarExpc = "data",
  TimeBgn = as.POSIXct('2020-01-01',tz='GMT'),
  TimeEnd = as.POSIXct('2020-01-02',tz='GMT'),
  NumDayExpiMax = NA
)
def.ucrt.meas.mult(data=data,calSlct=calSlct)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.ucrt.fdas.rstc.poly}

\link[NEONprocIS.cal]{def.ucrt.fdas.volt.poly}

\link[NEONprocIS.base]{def.log.init}

\link[NEONprocIS.cal]{wrap.ucrt.dp0p}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{L0'}
\keyword{calibration,}
\keyword{uncertainty,}
