% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.validate.info.cal.R
\name{def.validate.info.cal}
\alias{def.validate.info.cal}
\title{Validate calibration information returned by NEONprocIS.cal::def.read.cal.xml}
\usage{
def.validate.info.cal(infoCal, NameList = c("cal", "ucrt"),
  CoefCal = NULL, CoefUcrt = NULL, log = NULL)
}
\arguments{
\item{infoCal}{List of calibration and uncertainty information read from a NEON calibration file
(as from NEONprocIS.cal::def.read.cal.xml).}

\item{NameList}{A character vector of names of the minimum list elements in infoCal. Defaults to 
"cal" and "ucrt".}

\item{CoefCal}{A character vector of coefficient names expected to be present in the "Name" column
of data frame infoCal$cal. Defaults to NULL, which will not check for any expected coefficients.}

\item{CoefUcrt}{A character vector of coefficient names expected to be present in the "Name" column
of data frame infoCal$ucrt. Defaults to NULL, which will not check for any expected coefficients.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
Boolean. True if passes all checks. FALSE otherwise. \cr
}
\description{
Validate the list elements and coefficients expected in the cal and/or ucrt data frames returned
from NEONprocIS.cal::def.read.cal.xml. Returns True if passes all checks. FALSE otherwise.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
