% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.cal.conv.dp0p.R
\name{wrap.cal.conv.dp0p}
\alias{wrap.cal.conv.dp0p}
\title{Wrapper for applying calibration conversion to NEON L0 data}
\usage{
wrap.cal.conv.dp0p(data, calSlct, FuncConv, Meta = list(), log = NULL)
}
\arguments{
\item{data}{Data frame of L0 data. Must include POSIXct time variable readout_time.}

\item{calSlct}{A named list of data frames, list element corresponding to calibrated terms.
The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, as returned 
from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{FuncConv}{A data frame indicating the calibration functions to apply and (optionally) the 
L0 terms to apply them to. The columns of the data frame are:
\code{FuncConv}: Character. The calibration conversion function within the NEONprocIS.cal package . Note that 
any and all calibration functions specified here must accept arguments "data", "infoCal", "varCal", "slctCal", 
"Meta", and "log", even if they are unused in the function. See any def.cal.conv.____.R 
\code{var}: Character. The name of the variable/term to be calibrated. Typically this will be a single L0 term matching
a column in the input data frame. However, it can be a term not found in the input data frame, multiple terms separated 
by pipes (e.g. "resistance|voltage") or no term at all (indicated by an NA). These uncommon cases are acceptable so long 
as the calibration conversion function is able to handle the case, for example if multiple L0 terms are used to create 
a single calibrated output. \cr}

\item{Meta}{(optional). A named list (default is an empty list) containing additional metadata to pass to 
calibration and uncertainty functions. This can contain whatever information might be needed in the
calibration and/or uncertainty functions in addition to calibration and uncertainty information.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of the converted (calibrated) L0' data after running through the 
calibration functions listed in FuncConv. Note that the input data frame \code{data} 
is passed into each function in sequence, where the input to each subsequent function
is the data frame output by the previous function.
}
\description{
Wrapper function. Apply calibration conversion function to NEON L0 data, thus generating NEON
L0' data.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.conv.poly}

\link[NEONprocIS.cal]{def.cal.func.poly}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
