% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.cal.conv.R
\name{wrap.cal.conv}
\alias{wrap.cal.conv}
\title{Wrapper for applying calibration conversion to NEON L0 data}
\usage{
wrap.cal.conv(data, calSlct, FuncConv, DirCal = "./", log = NULL)
}
\arguments{
\item{data}{Data frame of L0 data. Must include POSIXct time variable readout_time.}

\item{calSlct}{A named list of data frames, list element corresponding to the variable for which
uncertainty coefficients are to be compiled. The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, as returned 
from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{FuncConv}{A data frame of the variables for to convert and the function to convert 
them with. Columns include:\cr
\code{var} Character. The variable in data for which to compute uncertainty \cr
\code{FuncConv} A character string indicating the calibration conversion function  
within the NEONprocIS.cal package that should be used. For most NEON data products, this will be 
"def.cal.conv.poly". Note that any alternative function must accept arguments "data", "infoCal", 
and "log", even if they are unused in the function, and must return a vector of converted data.
Input "data" to the function is an array of the data for the applicable term. Input "infoCal" is
a data frame of calibration information (including uncertainty) as returned from 
NEONprocIS.cal::def.read.cal.xml. If no calibration files are associated with the term, infoCal 
would be passed in to the function as NULL. Input "log" is a logger object as generated by 
NEONprocIS.base::def.log.init and used in this script to generate hierarchical logging.\cr}

\item{DirCal}{Character string. Relative or absolute path (minus file name) to the main calibration
directory. Nested within this directory are directories for each variable in calSlct, each holding
calibration files for that variable. Defaults to "./"}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of the converted (calibrated) L0' data, limited to the variables in FuncConv.
}
\description{
Wrapper function. Apply calibration conversion function to NEON L0 data, thus generating NEON
L0' data.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.conv.poly}

\link[NEONprocIS.cal]{def.cal.func.poly}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
