% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.cal.slct.R
\name{wrap.cal.slct}
\alias{wrap.cal.slct}
\title{Wrapper for selecting the applicable calibrations and their time ranges for all variables}
\usage{
wrap.cal.slct(DirCal = "./", TimeBgn, TimeEnd, NumDayExpiMax,
  log = NULL)
}
\arguments{
\item{DirCal}{Character string. Relative or absolute path (minus file name) to the main calibration
directory. Nested within this directory are directories for each variable, each holding
calibration files for that variable. Defaults to "./"}

\item{TimeBgn}{A POSIXct timestamp of the start date of interest (inclusive)}

\item{TimeEnd}{A POSIXct timestamp of the end date of interest (exclusive)}

\item{NumDayExpiMax}{A data frame indicating the max days since expiration that calibration 
information is still considered usable for each variable. Calibrations beyond this allowance period
are treated as if they do not exist. Columns in this data frame are:\cr
\code{var} Character. Variable name.\cr
\code{NumDayExpiMax} Numeric. Max days after expiration that a calibration is considered usable.\cr}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A named list, each element corresponding to the variables found in the DirCal directory
and holding a data frame of selected calibrations for the time range of interst as output by 
NEONprocIS.cal::def.cal.slct. See that function for details.
}
\description{
Wrapper function. Select the calibrations and their time ranges that apply for each 
variables for which calibration information is supplied.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.meta}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
