% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.ucrt.coef.R
\name{wrap.ucrt.coef}
\alias{wrap.ucrt.coef}
\title{Wrapper for compiling uncertainty coefficients for all variables and time ranges}
\usage{
wrap.ucrt.coef(
  calSlct,
  DirCal = "./",
  ParaUcrt = NULL,
  ucrtCoefFdas = NULL,
  mappNameVar = NULL,
  log = NULL
)
}
\arguments{
\item{calSlct}{A named list of data frames, list element corresponding to the variable for which
uncertainty coefficients are to be compiled. The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, as returned 
from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{DirCal}{Character string. Relative or absolute path (minus file name) to the main calibration
directory. Nested within this directory are directories for each variable in calSlct, each holding
calibration files for that variable. Defaults to "./"}

\item{ParaUcrt}{A data frame indicating which (if any) type of FDAS uncertainty coefficients 
apply to each variable. Columns must include:\cr
\code{var} Character. The variable/term name (same as those in calSlct)\cr
\code{typeFdas} A single character indicating the type of measurement made by NEON's field 
data acquisision system (GRAPE). Acceptable values are "R" for resistance measurements, "V" 
for voltage measurements, or NA (no quotes) for measurements in which FDAS uncertainty does
not apply (e.g. digital L0 output). \cr
A NULL entry for ParaUcrt (default) or variables missing from ParaUcrt indicate that FDAS 
uncertainty does not apply.}

\item{ucrtCoefFdas}{A data frame of FDAS uncertainty coefficients, as read by 
NEONprocIS.cal::def.read.ucrt.coef.fdas. Columns include:\cr
\code{Name} Character. Name of the coefficient.\cr
\code{Value} Character. Value of the coefficient.\cr
\code{.attrs} Character. Relevant attribute (i.e. units)\cr
Defaults to NULL, in which case no variables in ParaUcrt may indicate that FDAS uncertainty 
applies.}

\item{mappNameVar}{A data frame with in/out variable name mapping as produced by 
NEONprocIS.base::def.var.mapp.in.out. See documentation for that function. If input (default is
NULL), input variable names in the output data frames will be replaced by their corresponding 
output name.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A named list, each element corresponding to those in ParaUcrt$var and holding a data
frame of uncertainty coefficients and applicable time ranges.  \cr
}
\description{
Wrapper function. Compute individual measurement uncertainty for calibrated data at native
frequency (NEON L0' data).
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.base]{def.var.mapp.in.out}

\link[NEONprocIS.cal]{def.read.ucrt.coef.fdas}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
