% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.ucrt.dp0p.R
\name{wrap.ucrt.dp0p}
\alias{wrap.ucrt.dp0p}
\title{Wrapper for computing individual measurement uncertainty for NEON L0' data}
\usage{
wrap.ucrt.dp0p(data, FuncUcrt, calSlct, Meta = list(), log = NULL)
}
\arguments{
\item{data}{Data frame of L0 data. Must include POSIXct time variable readout_time.}

\item{FuncUcrt}{A data frame of the functions and variables for which individual measurement 
and/or FDAS uncertainty is to be calculated. Columns include:\cr
\code{FuncUcrt} A character string indicating the individual measurement (calibration) or FDAS 
uncertainty function within the NEONprocIS.cal package. For most NEON data products, 
this will be "def.ucrt.meas.cnst" or "def.ucrt.meas.mult" for measurement/calibration 
uncertainty, and "def.ucrt.fdas.rstc.poly" or "def.ucrt.fdas.volt.poly" for FDAS 
(data acquisition system) uncertainty. Note that any alternative function must accept 
the same arguments as these functions, even if they are unused, and return the same 
output format. See one of those functions for details. \cr
\code{var} Character. The variable(s) in input data frame 'data' that will be used in the 
uncertainty function specified in FuncUcrt. In most cases, this will be a single L0 
variable for which to compute uncertainty, but it can be any character string so long 
as the specified (custom) uncertainty function knows what to do with it. Note that the 
uncertainty function is responsible for naming the output list containing 
uncertainty data frames for each variable, and that any overlap in the names across 
the output list will cause the uncertainty data frames to be combined (intentionally -
see return information). Thus, ensure that the column names of data frames for the 
same variable (list name) are unique. In the standard measurement and FDAS uncertainty functions, 
the output list names will match the name of the L0 variable specified in \code{var}.\cr}

\item{calSlct}{A named list of data frames, list elements typically corresponding to the variables in
FuncUcrt$var. The data frame in each list element holds information about the calibration files and 
time periods that apply to the variable, as returned from NEONprocIS.cal::def.cal.slct. 
See documentation for that function. Assign NULL to list elements (variables) for which calibration
information is not applicable.}

\item{Meta}{(optional). A named list (default is an empty list) containing additional metadata to pass to 
calibration and uncertainty functions. This can contain whatever information might be needed in the
calibration and/or uncertainty functions in addition to calibration and uncertainty information. 
Note that the standard fdas uncertainty functions require fdas uncertainty coefficients to be 
provided in Meta$ucrtCoefFdas.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A named list, each list element containing a data frame of uncertainty data for a single
variable matching the name of the list element. 
Note that each row in each uncertainty data frame corresponds to the times in 
data$readout_time, but the variable readout_time is and should not be included in the output. 
The data frames for list elements with the same name will be combined, all columns
beginning with either ucrtMeas or ucrtFdas across the data frames for that variable
will be combined in quadrature and placed into a column labeled ucrtComb.
Each output uncertainty data frame will also includes a column named ucrtExpn,
which is the expanded (95%) uncertainty of ucrtComb (currently just multiplied by 2).
}
\description{
Wrapper function. Compute individual measurement and/or FDAS uncertainty for 
calibrated data at native frequency (NEON L0' data).
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.ucrt.meas.cnst}

\link[NEONprocIS.cal]{def.ucrt.meas.mult}

\link[NEONprocIS.cal]{def.ucrt.fdas.rstc.poly}

\link[NEONprocIS.cal]{def.ucrt.fdas.volt.poly}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
