% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.meas.rh.dew.frst.pt.R
\name{def.ucrt.meas.rh.dew.frst.pt}
\alias{def.ucrt.meas.rh.dew.frst.pt}
\title{Uncertainty for dew/frost point from the relative humidity sensor as part of the relative
humidity transition}
\usage{
def.ucrt.meas.rh.dew.frst.pt(
  data = data.frame(data = base::numeric(0)),
  infoCal = NULL,
  varUcrt = "dew_point",
  calSlct = NULL,
  log = NULL
)
}
\arguments{
\item{data}{Temperature, relative humidity, and dew/frost point data from the relative humidity sensor}

\item{infoCal}{Not used in this function \cr}

\item{varUcrt}{A character string of the target variable (column) in the data frame \code{data} for 
which uncertainty data will be computed (dew_point in this function).}

\item{calSlct}{Defaults to NULL. See the inputs to NEONprocIS.cal::wrap.ucrt.dp0p for what this input is.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A data frame with the following variables:\cr
\code{ucrtMeas_dew_point} - combined measurement uncertainty for an individual dew/frost point reading. Includes the
repeatability and reproducibility of the sensor and the lab DAS and uncertainty of the
calibration procedures and coefficients including uncertainty in the standard (truth).
}
\description{
Definition function. Alternative calibration uncertainty function. Create file (dataframe) with
uncertainty information based off of the L0 temperature, relative humidity, and dew/frost point data values from 
the relative humidity sensor according to NEON.DOC.000851 - NEON Algorithm Theoretical Basis   
Document (ATBD): Humidity and Temperature Sensor.

Accepts L0 data and NEON uncertainty information as produced
by NEONprocIS.cal::def.read.cal.xml and returns a vector of individual measurement
uncertainties for each data value.
}
\examples{
data <- data.frame(dew_point=c(1,-1,5,4,4.5),temperature=c(2,3,6,8,5),relative_humidity=c(1,6,7,0,10),
readout_time=as.POSIXct(c('2019-01-01 02:00','2019-01-01 04:01','2019-01-01 06:02','2019-01-01 08:01','2019-01-01 10:02'),tz='GMT'))
calSlct=list("temperature"= data.frame(timeBgn=as.POSIXct("2019-01-01",tz="GMT"),
timeEnd=as.POSIXct("2019-01-02",tz="GMT"),file = "30000000000080_WO29705_157555.xml",id = 157555, expi= FALSE),
"relative_humidity"= data.frame(timeBgn=as.POSIXct("2019-01-01",tz="GMT"),
timeEnd=as.POSIXct("2019-01-02",tz="GMT"),file = "30000000000080_WO29705_157554.xml",id = 157554, expi= FALSE),
"dew_point"= data.frame(timeBgn=as.POSIXct("2019-01-01",tz="GMT"),timeEnd=as.POSIXct("2019-01-02",tz="GMT"),
file = "30000000000080_WO29705_157556.xml",id = 157556, expi= FALSE))
def.ucrt.meas.rh.dew.frst.pt(data=data,calSlct=calSlct)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
NEON.DOC.000851 NEON Algorithm Theoretical Basis Document (ATBD): Humidity and Temperature Sensor
}
\seealso{
\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.ucrt.fdas.rstc.poly}

\link[NEONprocIS.cal]{def.ucrt.fdas.volt.poly}

\link[NEONprocIS.base]{def.log.init}

\link[NEONprocIS.cal]{wrap.ucrt.dp0p}
}
\author{
Edward Ayres \email{eayres@battelleecology.org}
}
\keyword{L0',}
\keyword{calibration,}
\keyword{dew}
\keyword{frost}
\keyword{hmp155,}
\keyword{point}
\keyword{point,}
\keyword{uncertainty,}
