% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.wq.do.R, R/def.ucrt.wq.do.conc.R
\name{def.ucrt.wq.do.conc}
\alias{def.ucrt.wq.do.conc}
\title{Uncertainty for dissolved oxygen (DO) concentration (mg/L) as part of the water
quality transition}
\usage{
def.ucrt.wq.do.conc(data, infoCal = NULL, log = NULL)

def.ucrt.wq.do.conc(data, infoCal = NULL, log = NULL)
}
\arguments{
\item{data}{Dissolved oxygen (DO) concentration data [vector]}

\item{infoCal}{List of calibration and uncertainty information read from a NEON calibration file
(as from NEONprocIS.cal::def.read.cal.xml). Included in this list must be infoCal$ucrt, which is
a data frame of uncertainty coefficents. Columns of this data frame are:\cr
\code{Name} String. The name of the coefficient. \cr
\code{Value} String or numeric. Coefficient value. Will be converted to numeric. \cr}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}

\item{data}{Dissolved oxygen (DO) concentration data [vector]}
}
\value{
dataframe with L0 uncertatinty column(s) [dataframe]

dataframe with L0 uncertatinty column(s) [dataframe]
}
\description{
Alternative calibration uncertainty function. Create file (dataframe) with 
uncertainty information based off of the L0 dissolved oxygen (DO) concentration data values 
according to NEON.DOC.004931 - NEON Algorithm Theoretical Basis Document (ATBD): Water Quality.

Note: This script implements logging described in \code{\link[NEONprocIS.base]{def.log.init}},
which uses system environment variables if available.

Alternative calibration uncertainty function. Create file (dataframe) with 
uncertainty information based off of the L0 dissolved oxygen (DO) concentration data values 
according to NEON.DOC.004931 - NEON Algorithm Theoretical Basis Document (ATBD): Water Quality.

Note: This script implements logging described in \code{\link[NEONprocIS.base]{def.log.init}},
which uses system environment variables if available.
}
\examples{
#Written to potentially plug in to def.cal.conv.R
ucrt <- def.ucrt.wq.do.conc(data = data, cal = NULL)
#Written to potentially plug in to def.cal.conv.R
ucrt <- def.ucrt.wq.do.conc(data = data, cal = NULL)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007

License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
None currently

None currently
}
\author{
Kaelin Cawley \email{kcawley@battelleecology.org}

Kaelin Cawley \email{kcawley@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
