% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.cal.conv.R
\name{wrap.cal.conv}
\alias{wrap.cal.conv}
\title{Wrapper for applying calibration conversion to NEON L0 data}
\usage{
wrap.cal.conv(data, calSlct, FuncConv, log = NULL)
}
\arguments{
\item{data}{Data frame of L0 data. Must include POSIXct time variable readout_time.}

\item{calSlct}{A named list of data frames, list element corresponding to calibrated terms.
The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, as returned 
from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{FuncConv}{A data frame of the terms/variables to convert and the function to convert 
them with. Columns include:\cr
\code{var} Character. The variable in data to apply calibration to. If this variable does not 
exist in the data, it must be created by the associated calibration function in FuncConv. \cr
\code{FuncConv} A character string indicating the calibration conversion function  
within the NEONprocIS.cal package that should be used. For most NEON data products, this will be 
"def.cal.conv.poly". Note that any alternative function must accept the same arguments as 
def.cal.conv.poly, even if they are unused. See that function for details.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of the converted (calibrated) L0' data, limited to the variables in FuncConv.
}
\description{
Wrapper function. Apply calibration conversion function to NEON L0 data, thus generating NEON
L0' data.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.cal.conv.poly}

\link[NEONprocIS.cal]{def.cal.func.poly}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
