% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.qf.cal.R
\name{wrap.qf.cal}
\alias{wrap.qf.cal}
\title{Wrapper for computing calibration flags for all variables and time ranges}
\usage{
wrap.qf.cal(data, calSlct, mappNameVar = NULL, log = NULL)
}
\arguments{
\item{data}{Data frame of L0 data. Must include POSIXct time variable readout_time.}

\item{calSlct}{A named list of data frames, list element corresponding to the variable for which
uncertainty coefficients are to be compiled. The data frame in each list element holds 
information about the calibration files and time periods that apply to the variable, as returned 
from NEONprocIS.cal::def.cal.slct. See documentation for that function.}

\item{mappNameVar}{A data frame with in/out variable name mapping as produced by 
NEONprocIS.base::def.var.mapp.in.out. See documentation for that function.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A named list of qfExpi and qfSusp, each holding data frames with the same dimension as data, 
with the exception that the variable readout_time is removed.\cr
\code{qfExpi} Integer. The expired/valid calibration flag. 0 = valid, non expired calibration available; 
1 = no calibration or expired calibration available. \cr
\code{qfSusp} Integer. The suspect calibration flag. 0 = calibration not suspect, 1 = calibration suspect, 
-1 = no cal to evaluate
}
\description{
Wrapper function. Compute valid calibration and suspect calibration flags for NEON L0 data.
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.cal]{def.qf.cal.susp}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
