% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.ucrt.dp0p.R
\name{wrap.ucrt.dp0p}
\alias{wrap.ucrt.dp0p}
\title{Wrapper for computing individual measurement uncertainty for NEON L0' data}
\usage{
wrap.ucrt.dp0p(data, ParaUcrt, ucrtCoefFdas = NULL, calSlct,
  DirCal = "./", mappNameVar = NULL, log = NULL)
}
\arguments{
\item{data}{Data frame of L0 data. Must include POSIXct time variable readout_time.}

\item{ParaUcrt}{A data frame of the variables for which individual measurement uncertainty is
to be calculated. Columns include:\cr
\code{var} Character. The variable in data for which to compute uncertainty \cr
\code{typeFdas} A single character indicating the type of measurement made by NEON's field 
data acquisision system (GRAPE). Acceptable values are "R" for resistance measurements, "V" 
for voltage measurements, or NA (no quotes) for measurements in which FDAS uncertainty does
not apply (e.g. digital L0 output) \cr
\code{FuncUcrt} A character string indicating the individual measurement (calibration) uncertainty 
function within the NEONprocIS.cal package that should be used. Note that this does not include 
FDAS uncertainty. For most NEON data products, this will be "def.ucrt.meas.cnst". Note that any 
alternative function must accept arguments "data", "infoCal", and "log", even if they are unused 
in the function, and must return a data frame, one column of which is labeled "ucrtMeas" which 
corresponds to the individual measurement uncertainty for the variable. Input "data" to the function
is an array of the data for the applicable term. Input "infoCal" is a data frame of calibration 
information (including uncertainty) as returned from NEONprocIS.cal::def.read.cal.xml. 
If no calibration files are associated with the term, infoCal would be passed in to the function as 
NULL. Input "log" is a logger object as generated by NEONprocIS.base::def.log.init and used in 
this script to generate hierarchical logging.\cr}

\item{ucrtCoefFdas}{A data frame of FDAS uncertainty coefficients, as read by 
NEONprocIS.cal::def.read.ucrt.coef.fdas. Columns include:\cr
\code{Name} Character. Name of the coefficient.\cr
\code{Value} Character. Value of the coefficient.\cr
\code{.attrs} Character. Relevant attribute (i.e. units)\cr
Defaults to NULL, in which case no variables in ParaUcrt may indicate that FDAS uncertainty 
applies.}

\item{calSlct}{A named list of data frames, list element corresponding to the variables in
ParaUcrt. The data frame in each list element holds information about the calibration files and 
time periods that apply to the variable, as returned from NEONprocIS.cal::def.cal.slct. 
See documentation for that function. Assign NULL to list elements (variables) for which calibration
information is not applicable (i.e. a function other than def.ucrt.meas.cnst is used to compute its
uncertainty).}

\item{DirCal}{Character string. Relative or absolute path (minus file name) to the main calibration
directory. Nested within this directory are directories for each variable in calSlct, each holding
calibration files for that variable. Defaults to "./"}

\item{mappNameVar}{A data frame with in/out variable name mapping as produced by 
NEONprocIS.base::def.var.mapp.in.out. See documentation for that function. If input (default is NULL),
output variable names will be appended as prefixes to the column names in each output data frame.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A named list, each element corresponding to those in ParaUcrt$var and holding a data
frame of uncertainty data. Note that each row in each data frame corresponds to the times in 
data$readout_time, but the variable readout_time is not included in the output. One column
in each data frame is labeled ucrtComb, corresponding to the combined measurement uncertainty
of the individual measurements and FDAS (if applicable). If FDAS uncertainty does not apply, 
ucrtComb is simply a copy of ucrtMeas. \cr
}
\description{
Wrapper function. Compute individual measurement uncertainty for calibrated data at native
frequency (NEON L0' data).
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.base]{def.var.mapp.in.out}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
