% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.ucrt.dp0p.R
\name{wrap.ucrt.dp0p}
\alias{wrap.ucrt.dp0p}
\title{Wrapper for computing individual measurement uncertainty for NEON L0' data}
\usage{
wrap.ucrt.dp0p(
  data,
  FuncUcrt,
  ucrtCoefFdas = NULL,
  calSlct,
  mappNameVar = NULL,
  log = NULL
)
}
\arguments{
\item{data}{Data frame of L0 data. Must include POSIXct time variable readout_time.}

\item{FuncUcrt}{A data frame of the variables for which individual measurement uncertainty is
to be calculated. Columns include:\cr
\code{var} Character. The variable in data for which to compute uncertainty \cr
\code{FuncUcrtMeas} A character string indicating the individual measurement (calibration) uncertainty 
function within the NEONprocIS.cal package that should be used. Note that this does not include 
FDAS uncertainty. For most NEON data products, this will be "def.ucrt.meas.cnst". Note that any 
alternative function must accept the same arguments as def.ucrt.meas.cnst, even if they are unused, and 
return the same 
See that function for details. \cr
\code{FuncUcrtFdas} A character string indicating the FDAS uncertainty function within the NEONprocIS.cal 
package that should be used, if FDAS uncertainty applies. This field may be NA, which indicates that
FDAS uncertainty does not apply and will not be calculated (e.g. digital L0 output).}

\item{ucrtCoefFdas}{A data frame of FDAS uncertainty coefficients, as read by 
NEONprocIS.cal::def.read.ucrt.coef.fdas. Columns include:\cr
\code{Name} Character. Name of the coefficient.\cr
\code{Value} Character. Value of the coefficient.\cr
\code{.attrs} Character. Relevant attribute (i.e. units)\cr
Defaults to NULL, in which case no variables in FuncUcrt may indicate that FDAS uncertainty 
applies.}

\item{calSlct}{A named list of data frames, list element corresponding to the variables in
FuncUcrt. The data frame in each list element holds information about the calibration files and 
time periods that apply to the variable, as returned from NEONprocIS.cal::def.cal.slct. 
See documentation for that function. Assign NULL to list elements (variables) for which calibration
information is not applicable (i.e. a function other than def.ucrt.meas.cnst is used to compute its
uncertainty).}

\item{mappNameVar}{A data frame with in/out variable name mapping as produced by 
NEONprocIS.base::def.var.mapp.in.out. See documentation for that function. If input (default is NULL),
output variable names will be appended as prefixes to the column names in each output data frame.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A named list, each element corresponding to those in FuncUcrt$var and holding a data
frame of uncertainty data. Note that each row in each data frame corresponds to the times in 
data$readout_time, but the variable readout_time is not included in the output. One column
in each data frame is labeled ucrtComb, corresponding to the combined measurement uncertainty
of the individual measurements and FDAS (if applicable). If FDAS uncertainty does not apply, 
ucrtComb is simply a copy of ucrtMeas. \cr
}
\description{
Wrapper function. Compute individual measurement uncertainty for calibrated data at native
frequency (NEON L0' data).
}
\examples{
Currently none
}
\references{
Currently none
}
\seealso{
\link[NEONprocIS.cal]{def.cal.slct}

\link[NEONprocIS.cal]{def.read.cal.xml}

\link[NEONprocIS.base]{def.var.mapp.in.out}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
